fix utest
