+++ python/paddle/fluid/tests/unittests/CMakeLists.txt
@@ -159,7 +159,6 @@ if ((NOT WITH_GPU) AND (NOT WITH_ROCM))
     LIST(REMOVE_ITEM TEST_OPS test_parallel_dygraph_sparse_embedding_over_height)
     LIST(REMOVE_ITEM TEST_OPS test_parallel_dygraph_transformer)
     LIST(REMOVE_ITEM TEST_OPS test_parallel_dygraph_sync_batch_norm)
-    list(REMOVE_ITEM TEST_OPS test_parallel_dygraph_gan)
     list(REMOVE_ITEM TEST_OPS test_parallel_dygraph_control_flow)
     LIST(REMOVE_ITEM TEST_OPS test_imperative_auto_mixed_precision)
     LIST(REMOVE_ITEM TEST_OPS test_fleet_base_single)
@@ -826,7 +825,6 @@ if(WITH_DISTRIBUTE AND WITH_GPU AND WITH_NCCL)
         set_tests_properties(test_parallel_dygraph_sparse_embedding PROPERTIES TIMEOUT 120)
         set_tests_properties(test_parallel_dygraph_transformer PROPERTIES TIMEOUT 120)
         set_tests_properties(test_parallel_dygraph_unused_variables PROPERTIES TIMEOUT 120)
-        set_tests_properties(test_parallel_dygraph_gan PROPERTIES TIMEOUT 120)
         set_tests_properties(test_parallel_dygraph_control_flow PROPERTIES TIMEOUT 120)
     endif()
 endif()
