+++ python/paddle/fluid/tests/unittests/test_parallel_dygraph_control_flow.py
@@ -21,37 +21,39 @@ import unittest
 import paddle.fluid as fluid
 from test_dist_base import TestDistBase
 from spawn_runner_base import TestDistSpawnRunner
-from parallel_dygraph_control_flow import TestControlFlow
 
 flag_name = os.path.splitext(__file__)[0]
 
-# class TestDygraphControlFlowSame(TestDistBase):
-#     def _setup_config(self):
-#         self._sync_mode = False
-#         self._nccl2_mode = True
-#         self._dygraph = True
 
-#     def test_net(self):
-#         if fluid.core.is_compiled_with_cuda():
-#             self.check_with_place(
-#                 "parallel_dygraph_control_flow_same.py",
-#                 delta=1e-5,
-#                 check_error_log=True,
-#                 log_name=flag_name)
+class TestDygraphControlFlowSame(TestDistBase):
+    def _setup_config(self):
+        self._sync_mode = False
+        self._nccl2_mode = True
+        self._dygraph = True
 
-# class TestFleetDygraphControlFlowSame(TestDygraphControlFlowSame):
-#     def _setup_config(self):
-#         self._sync_mode = False
-#         self._nccl2_mode = True
-#         self._dygraph = True
-#         self._use_fleet_api = True
+    def test_net(self):
+        if fluid.core.is_compiled_with_cuda():
+            self.check_with_place(
+                "parallel_dygraph_control_flow_same.py",
+                delta=1e-5,
+                check_error_log=True,
+                log_name=flag_name)
 
-# class TestFleetDygraphControlFlowSameAccGrad(TestDygraphControlFlowSame):
-#     def _setup_config(self):
-#         self._sync_mode = False
-#         self._nccl2_mode = True
-#         self._dygraph = True
-#         self._accumulate_gradient = True
+
+class TestFleetDygraphControlFlowSame(TestDygraphControlFlowSame):
+    def _setup_config(self):
+        self._sync_mode = False
+        self._nccl2_mode = True
+        self._dygraph = True
+        self._use_fleet_api = True
+
+
+class TestFleetDygraphControlFlowSameAccGrad(TestDygraphControlFlowSame):
+    def _setup_config(self):
+        self._sync_mode = False
+        self._nccl2_mode = True
+        self._dygraph = True
+        self._accumulate_gradient = True
 
 
 class TestDygraphControlFlowDiff(TestDistBase):
@@ -69,78 +71,21 @@ class TestDygraphControlFlowDiff(TestDistBase):
                 log_name=flag_name)
 
 
-# class TestFleetDygraphControlFlowDiff(TestDygraphControlFlowDiff):
-#     def _setup_config(self):
-#         self._sync_mode = False
-#         self._nccl2_mode = True
-#         self._dygraph = True
-#         self._use_fleet_api = True
+class TestFleetDygraphControlFlowDiff(TestDygraphControlFlowDiff):
+    def _setup_config(self):
+        self._sync_mode = False
+        self._nccl2_mode = True
+        self._dygraph = True
+        self._use_fleet_api = True
 
-# class TestFleetDygraphControlFlowDiffAccGrad(TestDygraphControlFlowDiff):
-#     def _setup_config(self):
-#         self._sync_mode = False
-#         self._nccl2_mode = True
-#         self._dygraph = True
-#         self._accumulate_gradient = True
 
-#     # def test_mnist(self):
-#     #     if fluid.core.is_compiled_with_cuda():
-#     #         self.check_with_place(
-#     #             "parallel_dygraph_control_flow.py",
-#     #             delta=1e-5,
-#     #             check_error_log=True,
-#     #             log_name=flag_name)
+class TestFleetDygraphControlFlowDiffAccGrad(TestDygraphControlFlowDiff):
+    def _setup_config(self):
+        self._sync_mode = False
+        self._nccl2_mode = True
+        self._dygraph = True
+        self._accumulate_gradient = True
 
-# class TestParallelDygraphControlFlow(TestDistBase):
-#     def _setup_config(self):
-#         self._sync_mode = False
-#         self._nccl2_mode = True
-#         self._dygraph = True
-
-#     def test_mnist(self):
-#         if fluid.core.is_compiled_with_cuda():
-#             self.check_with_place(
-#                 "parallel_dygraph_control_flow.py",
-#                 delta=1e-5,
-#                 check_error_log=True,
-#                 log_name=flag_name)
-
-# class TestControlFlowSpawn(TestDistSpawnRunner):
-#     def test_mnist_with_spawn(self):
-#         if fluid.core.is_compiled_with_cuda() and sys.version_info >= (3, 4):
-#             self.check_dist_result_with_spawn(
-#                 test_class=TestControlFlow, delta=1e-5)
-
-# class TestFleetDygraphControlFlow(TestDistBase):
-#     def _setup_config(self):
-#         self._sync_mode = False
-#         self._nccl2_mode = True
-#         self._dygraph = True
-#         self._use_fleet_api = True
-
-#     def test_mnist(self):
-#         if fluid.core.is_compiled_with_cuda():
-#             self.check_with_place(
-#                 "parallel_dygraph_control_flow.py",
-#                 delta=1e-5,
-#                 check_error_log=True,
-#                 log_name=flag_name)
-
-# class TestFleetDygraphControlFlowAccGrad(TestDistBase):
-#     def _setup_config(self):
-#         self._sync_mode = False
-#         self._nccl2_mode = True
-#         self._dygraph = True
-#         self._use_fleet_api = True
-#         self._accumulate_gradient = True
-
-#     def test_mnist(self):
-#         if fluid.core.is_compiled_with_cuda():
-#             self.check_with_place(
-#                 "parallel_dygraph_control_flow.py",
-#                 delta=1e-5,
-#                 check_error_log=True,
-#                 log_name=flag_name)
 
 if __name__ == "__main__":
     unittest.main()
